requirements.txt
discord.py
yt-dlp
ffmpeg-python
configparser
youtube-dl
spotipy
